Update index.js
lets say I have component A that uses RNAI. Currently, the code assumes the `data` are shown only after user does some interaction in A (eters text) and the `data` is changed, which will cause new `data` prop to be passed to RNAI. In componentWillReceiveProps, RNAI will get the new `data` and render the listView. In my use case, I already have a pre-populated list and it makes sense to display it to the user right away - so I'm testing the data array size (same as `this._showResults(dataSource.getRowCount() > 0);` on line 62) right in the constructor.